requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
Jinja2
click
itsdangerous
MarkupSafe
grpcio
blinker
backoff

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging